rainforest: extract the test code into rf_test.c
Now just building rf_test.c will result in the test code being built.
The code was moved out as-is.
